Added auth support to be able to develop images without watermarks

+++ tests/utils/test_files_context.py
@@ -14,21 +14,28 @@ class FileBackupContext:
     and the original state of the directory is restored.
     """
 
-    def __init__(self, directory: str) -> None:
+    def __init__(self, path_to_backup: str) -> None:
         """
-        Initialize the FileBackupContext with the specified directory.
+        Initialize the FileBackupContext with the specified directory or file path.
 
         Parameters
         ----------
-        directory : str
-            The path to the directory where files will be backed up and restored.
+        path_to_backup: str
+            The path to the directory or file to be backed up and restored.
 
         """
-        self.directory = directory
-        self.backup_dir = os.path.join(directory, f"backup_{int(time.time())}")
+        self.path_to_backup = Path(path_to_backup)
         self.files_to_backup = []
+        self.backup_path = self._get_backup_path()
 
-    def backup_files(self) -> None:
+    def _get_backup_path(self) -> Path:
+        """Determine the backup path based on whether it's a file or directory."""
+        backup_name = f"backup_{int(time.time())}_{self.path_to_backup.name}"
+        if self.path_to_backup.is_dir():
+            return self.path_to_backup.parent / backup_name
+        return self.path_to_backup.with_name(backup_name)
+
+    def backup_directory(self) -> None:
         """
         Backs up files in the specified directory, excluding any existing backup directories.
 
@@ -36,12 +43,12 @@ class FileBackupContext:
         except those beginning with 'backup_'.
         It then removes the original files and directories to prepare for new operations.
         """
-        if Path(self.directory).exists() and len(os.listdir(self.directory)) != 0:
-            Path(self.backup_dir).mkdir(parents=True)
-            for file in os.listdir(self.directory):
+        if Path(self.path_to_backup).exists() and len(os.listdir(self.path_to_backup)) != 0:
+            Path(self.backup_path).mkdir(parents=True)
+            for file in os.listdir(self.path_to_backup):
                 if not file.startswith("backup_"):
-                    src_path = os.path.join(self.directory, file)
-                    backup_path = os.path.join(self.backup_dir, file)
+                    src_path = os.path.join(self.path_to_backup, file)
+                    backup_path = os.path.join(self.backup_path, file)
                     if Path(src_path).is_dir():
                         shutil.copytree(src_path, backup_path)
                         shutil.rmtree(src_path)
@@ -49,6 +56,18 @@ class FileBackupContext:
                         shutil.copy(src_path, backup_path)
                         Path(src_path).unlink()
                     self.files_to_backup.append((src_path, backup_path))
+
+    def backup_file(self) -> None:
+        """
+        Backs up single file.
+
+        This method creates a backup of the file in the specified path.
+        It then removes the original file to prepare for new operations.
+        """
+        if Path(self.path_to_backup).exists() and Path(self.path_to_backup).is_file():
+            shutil.copy(self.path_to_backup, self.backup_path  )
+            Path(self.path_to_backup).unlink()
+            self.files_to_backup.append((self.path_to_backup, self.backup_path))
 
     def restore_files(self) -> None:
         """
@@ -58,14 +77,15 @@ class FileBackupContext:
         and then restores the original files and directories from the backup.
         The backup directory is removed after the restoration.
         """
-        # Clear the directory before restoring files
-        for file in os.listdir(self.directory):
-            if not file.startswith("backup_"):
-                file_path = os.path.join(self.directory, file)
-                if Path(file_path).is_dir():
-                    shutil.rmtree(file_path)
-                else:
-                    Path(file_path).unlink()
+        if Path(self.path_to_backup).is_dir():
+            # Clear the directory before restoring files
+            for file in os.listdir(self.path_to_backup):
+                if not file.startswith("backup_"):
+                    file_path = os.path.join(self.path_to_backup, file)
+                    if Path(file_path).is_dir():
+                        shutil.rmtree(file_path)
+                    else:
+                        Path(file_path).unlink()
         # Restore files from the backup
         for src_path, backup_path in self.files_to_backup:
             if Path(backup_path).is_dir():
@@ -73,12 +93,15 @@ class FileBackupContext:
             else:
                 shutil.copy(backup_path, src_path)
         # Remove the backup directory
-        if Path(self.backup_dir).is_dir():
-            shutil.rmtree(self.backup_dir)
+        if self.backup_path is not None and Path(self.backup_path).is_dir():
+            shutil.rmtree(self.backup_path)
+        # Remove the backup file
+        if self.backup_path is not None and Path(self.backup_path).is_file():
+            Path(self.backup_path).unlink()
 
     def __enter__(self) -> "FileBackupContext":
         """
-        Enter the context manager, backing up files in the specified directory.
+        Enter the context manager, backing up files in the specified directory or single file.
 
         Returns
         -------
@@ -86,7 +109,10 @@ class FileBackupContext:
             The context manager instance with files backed up.
 
         """
-        self.backup_files()
+        if Path(self.path_to_backup).is_dir():
+            self.backup_directory()
+        if Path(self.path_to_backup).is_file():
+            self.backup_file()
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb) -> None:  # noqa: ANN001
